BREAKING: Refactoring of the javascript-interface
The old interface was just the interface of Customize
extended by some functions to add compatibility to old versions
of Bootprint. It was designed before Customize was extracted
from Bootprint.

The new interface is much simpler, thought not as flexible. It does not
support chaining module loads and configuration merges.
It just loads Customize with

* one module (loaded either from `bootprint-modulename` or from
`modulename`-directly)
* one merged configuration

This is what the CLI-interface supports and it is now also the scope
of the javascript-interface.

If you need to load multiple modules and merge multiple configurations,
you can either use Customize directly, or create a customize-module that
does all the loading and merging.