Re-add and refactor dev-mode and dev-server
- MultiWatcher-class for file-watching + Tests
- Fix "onlyEngine"-parameter (was wrongly named "onlyEngine")
- Update customize-write-files for 2.0.1 to fix usage with "onlyEngine"
- Work-around for chokidar for IntelliJ: Re-add changed and unlinked
  files shortly after the event (see below for details)
- Lots of tests

Why the dev-mode remains in Bootprint:

The reason for removing the dev-mode was to reduce the package size,
which is mostly important, if Bootprint is used as dependency
(e.g. in a build-environment).
The rationale now is that Bootprint mostly a command-line-interface
for Customize. Instead of using the JavaScript-API of Bootprint,
people should rather use Customize directly. The exact use is still to
be documented in this package, but it should be mostly straight-forward
and no more than a couple of lines of code, so it does not really
justify creating a separate package.

About chokidar and IntelliJ:

I have attempted to use another file-watcher library instead of chokidar
for the dev-mode. I have come across issues when editing files with
IntelliJ. Chokidar seems to loose track of changed files if they are
directly watched (not the parent directory).
As a (rather hacky workaround, the watch will now be recreated after
very customize-run. I have not been able to reproduce this in a
test-case.

Other libraries (filewatcher, watchr etc) either did not use deep-watching
like chokidar or seemed to fall back to polling very quickly, or they
that the same-problem as chokidar. Chokidar seems to be the best choice
at the moment and the work-around works well with IntelliJ. And the
unit-tests pass as well...